Move functions used to create the table from App.js
